pdp_copyleft_detector/requirements.txt
# torch>=2.1.0
# torchvision>=0.16.0
timm>=0.9.12
open_clip_torch>=2.23.0
faiss-cpu>=1.7.4
# numpy>=1.24.0
# pillow>=10.0.0
# opencv-python>=4.9.0.80
scikit-learn>=1.4.0
albumentations>=1.4.3
# pandas>=2.2.0
# tqdm>=4.66.0
# matplotlib>=3.8.0
kornia>=0.7.3
kornia_moons>=0.2.8
einops>=0.7.0
transformers>=4.41.0
sentence-transformers>=2.6.1
easyocr>=1.7.1
pyyaml>=6.0.1
